+++ src/platform_impl/windows/window.rs
@@ -1108,6 +1108,11 @@ where
         },
     };
 
+    if pl_attribs.tool_window {
+        window_flags.set(WindowFlags::TOOL, true);
+        window_flags.set(WindowFlags::ON_TASKBAR, false);
+    }
+
     let mut initdata = InitData {
         event_loop,
         attributes,